requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
openai==1.3.5
pymilvus==2.3.4
numpy==1.24.3
requests==2.31.0
pyngrok==6.0.0 